Use xorriso for ISO creation
Switch to using xorriso for ISO creation as it is the default tool
for the job on modern distributions. No changes for RHEL ISOs.

xorriso created Windows ISOs boot on UEFI but not on BIOS, however
since Windows 11 already requires UEFI this shouldn't be much of
an issue.

Thanks to Thomas Schmitt for helping with xorriso details.